==> Analyses/01_1_Marsbrain_Place_Patches_Automatically.Rmd <==
---
title: "Automatic_Patch_Placement"
author: "Vera Weisbecker"
date: "14 February 2020"
output: html_document
---


```{r}


# Load libraries
library(rgl) #version 0.100.54
library(geomorph) #version 3.3.1
library(abind) #version 1.4-5
library(plyr) #version 1.8.6
library(vegan) #version 2.5-6
library(ape) #version 5.4-1
library(geiger) #version 2.0.7
library(nlme) #version 3.1-149
library(caper)#version 1.0.1
library(Morpho)#version 2.8


```

#read ckpt based manual landmarks

```{r}


iter = 1000 # setting number of iterations for all of the phylogenetic analyses

# Read in classifiers and specimen volumes data 
ind.SpecData <- read.csv("../Data/Raw/Specimen_data.csv", header=T, row.names = 1 )

#For the sake of having everything in one spot, a lot of additional acquisition notes are in the Specimen.data file. Paring it down to what is required for this analysis
ind.SpecData <- ind.SpecData[1:13]

## define sliding semilandmarks on curves (curveslide), and patches (surfslide) 
curveslide=as.matrix(read.csv("../Data/Raw/curveslide_manual.csv", header=T))
surfslide=as.matrix(read.csv("../Data/Raw/surfslide_manual.csv",header=T))

####~~~~~~~~~~~~~~loading 3d coordinate files~~~~~~~~~~~~~~~~~####

#FIRST REPLICATE DATASET
filelist <- list.files(path= "../Data/Raw/Replicates_1", pattern = "*.txt")
names <- gsub (".txt", "", filelist) # extracts names of specimens from the file name
filelist <- paste("../Data/Raw/Replicates_1/", filelist, sep="") # rename with path
coords <- NULL # make empty object that will be filled with 3D array of coordinate data
for (i in 1:length(filelist)){
  temp <- read.morphologika(filelist[i])
  k <- dim(temp)[1]
  coords <- rbind(coords, two.d.array(temp)) }
rep.1 <- arrayspecs(coords, k, 3)
dimnames(rep.1)[[3]] <- names
remove(i, filelist, names, k, coords, temp) # clean up environment

##SECOND REPLICATE DATASET
filelist <- list.files(path= "../Data/Raw/Replicates_2", pattern = "*.txt")
names <- gsub (".txt", "", filelist) # extracts names of specimens from the file name
filelist <- paste("../Data/Raw/Replicates_2/", filelist, sep="") # rename with path
coords <- NULL # make empty object that will be filled with 3D array of coordinate data
for (i in 1:length(filelist)){
  temp <- read.morphologika(filelist[i])
  k <- dim(temp)[1]
  coords <- rbind(coords, two.d.array(temp)) }
rep.2 <- arrayspecs(coords, k, 3)
dimnames(rep.2)[[3]] <- names
remove(i, filelist, names, k, coords, temp) # clean up environment

Raw3D <- abind(rep.1,rep.2) # abind concatenates (stacks) two arrays
remove(rep.1,rep.2) # clean up environment


```


#A  number of files were exported from the CT scans 1000x larger than they actually were due to issues with unit attribution (Sometimes Mimics does not like um-scale imports). In checkpoint, this was dealt with by exporting the LMs in smaller units. For automatic placement of surface semilandmarks, the files that were too large were separated into a "Larger Scale" folder. Note that all scans with this scaling property were ascertained to have accurate volumes(i.e. the volumes in the mimics files are not out by 1000).



```{r}

#Ensuring all ply files, specimen data, and replicates have identical specimen names

filelist1 <- list.files(path= "../Data/Raw/Replicates_1", pattern = "*.txt")
filelist2 <- list.files(path= "../Data/Raw/Replicates_2", pattern = "*.txt")

#make a list of all the ply files used in this study (separated into larger and smaller endocast files), then order the list alphabetically

plylist <-  c(list.files(path= "../Data/Raw/plys/Larger_scale", pattern = "*.ply"), list.files(path= "../Data/Raw/plys/Base_scale", pattern = "*.ply")); plylist <-plylist[order(plylist)]


#Checks to make sure data match up 

#replicate names - needs to be TRUE
filelist1 == filelist2

#ply and replicate file names - output should be integer(0)
plyfilematch <- sub(".ply", "", plylist)==sub(".txt", "", filelist1)
which(plyfilematch=="FALSE")

# match up specimen data file with replicate list; output needs to be integer(0)
Classifier_3dmatch <- match (rownames(ind.SpecData),sub(".txt", "", filelist1))
which(is.na(Classifier_3dmatch))

#Needs to be TRUE

dimnames(Raw3D)[[3]]==sub(".txt", "", filelist2)



#Ply file and species names also need to be the same, needs to be TRUE 
length(plylist)== length (dimnames(Raw3D[,,1:84])[[3]])
  

```

#Find the mean specimen for just fixed and curve landmarks, read in ply

```{r}


#remove all patch landmarks

Fixed_plus_curve <- Raw3D[-surfslide[,2], ,]

#This is just to determine what the new lm numbers are for the curve
#plot3d(Fixed_plus_curve)
#text3d(Fixed_plus_curve[,,1], texts =  c(1:39))

curveslide_NoPatch <-as.matrix(read.csv("../Data/Raw/curveslide_after_patch_removal.csv", header=T))

#Finding the mean specimen and making sure that the csizes are OK
ForMeanSpec <- gpagen(Fixed_plus_curve,curves = curveslide_NoPatch )
pca<-gm.prcomp(ForMeanSpec$coords)

plot( pca$x[,1]~log(ForMeanSpec$Csize))
text(pca$x[,1]~log(ForMeanSpec$Csize), labels=dimnames(pca$x)[[1]])

findMeanSpec(ForMeanSpec$coords) 


#Read in mean shape, create atlas

Mean_shape <-read.ply(file="../Data/Raw/plys/Base_scale/Ptap_JM12395.ply")


Atlas <- createAtlas(Mean_shape, landmarks =  Raw3D[-surfslide[,2], ,64],patch=Raw3D[surfslide[,2], ,64]  )

#To check if the landmarks are placed as required:
#plotAtlas(atlas = Atlas)

```

#The below places patches on whatever files are in a ply folder. It here places the patch in the folder with most Plys, but it can also be used to place patches in sub-folders just by re-naming "folder"", which was done initially to check that all patch placements worked. Applying the patch takes a while! 
```{r}

#To subset the data into the files existing in a folder - makes it easier to place the patch on just some selected plys while testing batches of files

Folder <- "../Data/Raw/plys/Base_scale/"

Ply_folder_list <- sub(".ply", "", list.files(Folder))

#identify where the plys in the chosen folder sit in the coordinate array (as not all plys are in the one folder), pull out the numbers for later placing on the plys

matching <- match(dimnames(Raw3D)[[3]],Ply_folder_list)
for_patch_placing <-which (!is.na(matching)==TRUE)


#This should be 140 (70 specimens in the Base_scale folder). If less, check file spelling!
length(for_patch_placing)

data <-Raw3D[-surfslide[,2], , for_patch_placing]

dimnames(data)[[3]]

#Place all the patches. Be aware that when the landmark data file names don't match with the plys in the folder, you will not see an error message here! You just end up with fewer species.
Patched <- placePatch(Atlas, data, path = Folder, inflate = 5)

#Check if lm configuration works
#checkLM(Patched, path = Folder, atlas=Atlas)

#Are specimen 1 and specimen 71 (first specimen of each replicate) the same name? Needs to be TRUE
dimnames(Patched)[[3]][1]==dimnames(Patched)[[3]][71]

#Making sure that the species have different coordinates (i.e. not accidentally overwritten rep 1 by rep 2) - this needs to be FALSE

Patched[,,1]==Patched[,,71]

```
# The below is scaling the raw landmarks to be appropriately sized for the outsized sized meshes; they are then taken back to their appropriate sizes after patch placement.

#NOTE that the replicates have to be kept separate initially because otherwise the re-ordering step before GPA (next chunk) will overwrite rep1 with rep2.

```{r}

#To subset the data into the files existing in a folder - makes it easier to place the patch on just some selected plys while testing batches of files
Folder <- "../Data/Raw/plys/Larger_scale/"

Ply_folder_list <- sub(".ply", "", list.files(Folder))

#identify where the plys in the chosen folder sit in the coordinate array, pull out the numbers for later placing on the plys
matching <- match(dimnames(Raw3D)[[3]],Ply_folder_list)
for_patch_placing <-which (!is.na(matching)==TRUE)


#The below is where the re-scaling of coordinates occurs
data <-Raw3D[-surfslide[,2], , for_patch_placing]*1000


dimnames(data)[[3]]

Patched_Larger_Scale <- placePatch(Atlas, data, path = Folder, inflate = 5)

#It is obviously important to check that the landmark placement worked! Commented out here so the code re-runs faster

#checkLM(Patched_Larger_Scale, path = Folder, atlas=Atlas)


#Take the coordinates back to "base scale" (i.e. divide coords by 1000)
Patched_to_base <-Patched_Larger_Scale/1000


Patched_coords_all_rep1 <- abind (Patched[,,c(1:70)], Patched_to_base[,,1:14])

Patched_coords_all_rep2 <- abind (Patched[,,c(71:140)], Patched_to_base[,,15:28])

length(dimnames(Patched)[[3]])
length(dimnames(Patched_to_base)[[3]])


dimnames(Patched_coords_all_rep1)[[3]]==dimnames(Patched_coords_all_rep2)[[3]]

#Earlier I was missing one ply file, this is how you search for it
#Missing <- which (dimnames(Raw3D)[[3]] %in% dimnames(Patched_coords_all)[[3]]=="FALSE")
#dimnames(Raw3D)[[3]][Missing]



```



# Last step:re-order the names so they match the species list for rep1 and re2 - then bind the arrays together and do some final checks to ensure all works. 
```{r}

Automatic_Raw_rep1 <- Patched_coords_all_rep1[,,match(dimnames(Raw3D[,,c(1:84)])[[3]],dimnames(Patched_coords_all_rep1)[[3]])]
Automatic_Raw_rep2 <- Patched_coords_all_rep2[,,match(dimnames(Raw3D[,,c(85:168)])[[3]],dimnames(Patched_coords_all_rep2)[[3]])]

#this should be TRUE
dimnames(Automatic_Raw_rep1)[[3]]==dimnames(Automatic_Raw_rep2)[[3]]

#This should be FALSE
Automatic_Raw_rep1[,,1]  == Automatic_Raw_rep2[,,1]

#put all together
Automatic_Raw <-abind(Automatic_Raw_rep1, Automatic_Raw_rep2)



#Did the matching work? Should be TRUE

dimnames(Raw3D)[[3]] == dimnames(Automatic_Raw)[[3]]

#If the 1st and 2nd reps are bound together before the re-ordering according to match(), the corresponding specimens of one rep are overwritten with the other one, resulting in identical coordinates. The below has to return FALSE and the PCA needs to look similar to the no-patch PCA.
Automatic_Raw[,,1]==Automatic_Raw[,,85]

#Comparing PCAs:

Automatic_Raw_GPA <- gpagen(Automatic_Raw)

Automatic_Patched <- gm.prcomp(Automatic_Raw_GPA$coords)




plot( Automatic_Patched$x[,2]~Automatic_Patched$x[,1])
text(Automatic_Patched$x[,2]~Automatic_Patched$x[,1], labels=dimnames(Automatic_Patched$x)[[1]])

NoPatchGPA <- gpagen(Fixed_plus_curve)

NoPatched <- gm.prcomp(NoPatchGPA$coords)

plot( NoPatched$x[,2]~NoPatched$x[,1])
text(NoPatched$x[,2]~NoPatched$x[,1], labels=dimnames(NoPatched$x)[[1]])

#Same length
dimnames(NoPatched$x)[[1]]==dimnames(Automatic_Patched$x)[[1]]

# Plotting the two together for comparison:
par(mfrow=c(2,1))

plot( Automatic_Patched$x[,2]~Automatic_Patched$x[,1])
text(Automatic_Patched$x[,2]~Automatic_Patched$x[,1], labels=dimnames(Automatic_Patched$x)[[1]])

plot( NoPatched$x[,2]~NoPatched$x[,1])
text(NoPatched$x[,2]~NoPatched$x[,1], labels=dimnames(NoPatched$x)[[1]])

```



#Save all to file; Prior to saving, make a new "processed" folder

```{r}

save(Automatic_Raw, file = "../Data/Processed/Automatic_Patch_Placement_raw.rda")

```
