==> Analyses/03_Marsbrain_Figures.Rmd <==
---
title: "03_Marsbrain_Figures"
author: "Vera Weisbecker & Emma Sherratt"
date: "2 March 2020"
output: html_document
---

```{r}

# Load  libraries
library(png) #version 0.1-7
library(magick) # 2.4.0
library(Morpho) #version 2.8
library (geomorph) #version 3.3.1
library(landvR) #version 0.5
library(ape) #version 5.4-1

load("../Data/Processed/WeisbETal_Data.RData")




```

#Create reference mesh (specimen mesh warped to mean shape). THIS DOES NOT NEED TO BE RUN AGAIN AFTER THE RELEVANT PLYS HAVE BEEN LOADED AND WRITTEN, so it is commented out here


```{r}

# #This is the specimen that was used for the Atlas. It is close to the mean shape (it is the mean specimen of a gpa without patches and has the advantage of ventrally concave cerebral hemispheres, which are otherwise hard to warp ))
# 
# # Load in the PLY file, give colour information
# mesh <- read.ply(file = "../Data/Raw/plys/Ptap_JM12395.ply")
# mesh$material$color <- mesh$it
# mesh$material$color <- "pink"
# 
# mesh_coords <- Automatic_Raw[, , 64]
# 
# #Open a new 3d window because the colour does not update from read.ply. There is no problem with plotting this!
# open3d()
# shade3d(mesh) #
# points3d(mesh_coords, col = "green", size = 10)
# 
# # get the mean shape
# ref <- mshape(gdf_automatic$coords)
# 
# 
# #plotting landmarks on the specimen to ensure no mistakes were made. ptsize has to be set to very small and colour needs specifying as grey otherwise there are little points on the specimen representing the vertices.
# 
# ##but the landmarks don't show up UNLESS the lms are designated as all fixed. I don't know how to navigate to overall point size.
# 
# plotspec(
#   spec = mesh,
#   digitspec = mesh_coords,
#   col = "grey",
#   mesh.ptsize = 0.01,
#   centered = FALSE,
#   fixed = NULL,
#   fixed.pt.col = "green",
#   fixed.pt.size = 1000
# )
# 
# #warp the specimen mesh to the mean shape. THIS THROWS AN ERROR THAT MENTIONS "FIXED" SO I STRONGLY SUSPECT IT'S GOT SOMETHING TO DO WITH FIXED LANDMARK DESIGNATION CHANGES.
# refmesh <-
#   warpRefMesh(mesh, mesh_coords, ref, color = "pink", centered = FALSE)
# 
# #Also check if this all works
# shade3d(refmesh) #
# points3d(ref, col = "green", size = 10)
# 
# #To write the ply file, you need to run shade3d of the mesh first.
# open3d()
# 
# shade3d(refmesh)
# 
# writePLY("../Figures/3DModels/refmesh.ply", format = "ascii")


```

#Plotting preparation
```{r}

### Make clade-level colour vector
cols <- c("blue", "dark green", "#00CCCC", "#CC0000", "#0099FF", "light green", 
          "#9966FF", "#FF6600", "#FF6600", "#00CC00", "#FFCC00", "dark blue") 
  # specific colours for clades; can be hex codes or real names
  names(cols)=levels(as.factor(gdf_automatic$SpecData.Cladename))
  col.gp <- cols[match(gdf_automatic$SpecData.Cladename, names(cols))] 

## Make a pch vector for fossil versus non.fossil.
fos.gp <- rep("21", length(col.gp))
  fos.gp[which(gdf_automatic$SpecData.Fossil == "yes")] <- "8"
  fos.gp <- as.numeric(fos.gp)

# Make a short form vector for plotting species names, in G.sp formal
Gen.sp <- matrix(unlist(strsplit(as.character(gdf_automatic$SpecData.Species), "_")), 
                 ncol=2, byrow = T)
Gen.sp <- paste(substr(Gen.sp[,1], 1, 1), substr(Gen.sp[,2], 1, 3), sep=".")



#PCA setup
# preparing axis labels. 
PCA_summary <- summary(PCA)

pc1lab <-paste("PC 1 ","(",
               round(100*PCA_summary$PC.summary[2,1],1),"%)",sep="")
pc2lab <-paste("PC 2 ","(",
               round(100*PCA_summary$PC.summary[2,2],1),"%)",sep="")


### creating position matrices to use for looping through identical views of 3d representations; this does not need to be run again and should be in the published github repo

# FOV = 0 # sets parallel projection
# open3d(); view3d(fov=0);shade3d(refmesh) # open a mesh on rgl
# usrMat.dorsal <- par3d()$userMatrix # set by hand, adjust specimen into dorsal positon
#   write.csv(usrMat.dorsal, "usrMat.dorsal.csv",row.names=F)
# usrMat.ventral <- par3d()$userMatrix # set by hand, adjust specimen into ventral positon
#   write.csv(usrMat.ventral, "usrMat.ventral.csv",row.names=F)
# usrMat.lateral <- par3d()$userMatrix # set by hand, adjust specimen into lateral positon
#   write.csv(usrMat.lateral, "usrMat.lateral.csv",row.names=F)
# usrMat.posterior <- par3d()$userMatrix # set by hand, adjust specimen into posterior positon
#   write.csv(usrMat.posterior, "usrMat.posterior.csv",row.names=F)

## I've already set them so just import:
usrMat.dorsal <- as.matrix(read.csv("../Figures/usrMat.dorsal.csv", header = T))
usrMat.ventral <- as.matrix(read.csv("../Figures/usrMat.ventral.csv", header = T))
usrMat.lateral <- as.matrix(read.csv("../Figures/usrMat.lateral.csv", header = T))
usrMat.posterior <- as.matrix(read.csv("../Figures/usrMat.posterior.csv", header = T))

# ## Usage
# open3d(FOV=0, userMatrix = usrMat.dorsal, windowRect=windowRect) # for a dorsal view
# open3d(FOV=0, userMatrix = usrMat.lateral, windowRect=windowRect) # for a lateral view
# open3d(FOV=0, userMatrix = usrMat.posterior, windowRect=windowRect) # for a posterior view


#mean shape used throughout
ref <- mshape(gdf_automatic$coords)

#The code to produce this mesh is below but does not need to be run again.The addition of colour is a pain but the reading in of plys seems to require it (it may not on other computers and/or other ply files)

refmesh <- read.ply("..//Figures/3DModels/refmesh.ply")
refmesh$material$color <- refmesh$it
refmesh$material$color <- "pink"

# Partition the brain into its parts; this was done manually into a csv file but it is also possible to do this using the define.module function in geomorph

#part.gp: 1 = OB, 2 = Cerebrum, 3 = Cerebellum, 4 = Brain base

part.gp <- read.csv(file="../Data/Raw/partitions_automatic.csv", header = FALSE)

#Getting the coords of the same lm configuration that was also used for the atlas (there are 2 reps for each specimen and here we're taking the 1st)

mesh_coords <- Automatic_Raw[,,(which(dimnames(Automatic_Raw)[[3]]=="Ptap_JM12395"))[1]]

#colouring the landmarks

lmk.col <- rep("darkorange",dim(mesh_coords)[1])
  lmk.col[which(part.gp==2)] <- "red"
  lmk.col[which(part.gp==3)] <- "green"
  lmk.col[which(part.gp==4)] <- "purple"
  
lmk <- gridPar(pt.bg=lmk.col)


PC1min <- PCA$shapes$shapes.comp1$min
PC1max <- PCA$shapes$shapes.comp1$max
PC2min <- PCA$shapes$shapes.comp2$min
PC2max <- PCA$shapes$shapes.comp2$max



```



#Make coloured mesh with fixed and curve semilandmarks

```{r}



curve <- read.csv("../Data/Raw/curveslide_automatic.csv")
curve <- curve[,2]

surf <- read.csv("../Data/Raw/surfslide_automatic.csv")
surf <- surf$x

fixed <- as.vector(c(1:71))
fixed <- fixed [-c(which(fixed %in% curve),which(fixed %in% surf)) ]


#This code plots the brains in dorsal, lateral and ventral view together with the landmarks. This is done using par3d




open3d(FOV=0,windowRect=c(0,0,1000,3000), zoom=0.7 )


#set up image with three subscenes in one column
mfrow3d(3,1)

#plot dorsal view
shade3d(refmesh,tran = par3d(userMatrix=usrMat.dorsal ))
spheres3d(ref[fixed,1], ref[fixed,2], ref[fixed,3], col="purple", lit=TRUE,radius = 0.008) 
spheres3d(ref[surf,1], ref[surf,2], ref[surf,3], col="yellowgreen", lit=TRUE,radius = 0.005)     
spheres3d(ref[curve,1], ref[curve,2], ref[curve,3], col="turquoise", lit=TRUE,radius = 0.005)
#text3d(ref, texts = c(1:71), cex=2)

#move to next subscene
next3d()

#plot lateral view
shade3d(refmesh,tran = par3d(userMatrix=usrMat.lateral))
spheres3d(ref[fixed,1], ref[fixed,2], ref[fixed,3], col="purple", lit=TRUE,radius = 0.008) 
spheres3d(ref[surf,1], ref[surf,2], ref[surf,3], col="yellowgreen", lit=TRUE,radius = 0.005)     
spheres3d(ref[curve,1], ref[curve,2], ref[curve,3], col="turquoise", lit=TRUE,radius = 0.005) 

#move to next subscene
next3d()

#plot ventral view
shade3d(refmesh,tran = par3d(userMatrix=usrMat.ventral) )
spheres3d(ref[fixed,1], ref[fixed,2], ref[fixed,3], col="purple", lit=TRUE,radius = 0.008) 
spheres3d(ref[surf,1], ref[surf,2], ref[surf,3], col="yellowgreen", lit=TRUE,radius = 0.005)     
spheres3d(ref[curve,1], ref[curve,2], ref[curve,3], col="turquoise", lit=TRUE,radius = 0.005)   

rgl.snapshot(filename = "../Figures/FigX_Lm_protocol.png")

```

#Plot the landmark numbers for ventral and dorsal views; requires above landmark designations

```{r}


open3d(FOV=0,windowRect=c(0,0,1400,2100), zoom=0.6 )

Aux_curve <- c(44,47,50,53)

#shade3d(refmesh)
#text3d(ref, texts = c(1:71), cex=3)

#set up image with two subscenes in one column
mfrow3d(2,1)

next3d()
#Plot view with just fixed landmark numbers

shade3d(refmesh,tran = par3d(userMatrix=usrMat.dorsal ), alpha=0.4)
spheres3d(ref[fixed,1], ref[fixed,2], ref[fixed,3], col="purple", lit=TRUE,radius = 0.002) 
spheres3d(ref[surf,1], ref[surf,2], ref[surf,3], col="yellowgreen", lit=TRUE,radius = 0.002)     
spheres3d(ref[curve,1], ref[curve,2], ref[curve,3], col="turquoise", lit=TRUE,radius = 0.002)
text3d(ref[fixed,1], ref[fixed,2], ref[fixed,3], texts = fixed, cex=2.5 )
text3d(ref[Aux_curve,1], ref[Aux_curve,2], ref[Aux_curve,3], texts = Aux_curve, cex=2.5, col="royalblue3" )


next3d()
#Plot view with just fixed landmark numbers

shade3d(refmesh,tran = par3d(userMatrix=usrMat.ventral ), alpha=0.4)
spheres3d(ref[fixed,1], ref[fixed,2], ref[fixed,3], col="purple", lit=TRUE,radius = 0.002) 
spheres3d(ref[surf,1], ref[surf,2], ref[surf,3], col="yellowgreen", lit=TRUE,radius = 0.002)     
spheres3d(ref[curve,1], ref[curve,2], ref[curve,3], col="turquoise", lit=TRUE,radius = 0.002)
text3d(ref[fixed,1], ref[fixed,2], ref[fixed,3], texts = fixed, cex=2.5 )
text3d(ref[Aux_curve,1], ref[Aux_curve,2], ref[Aux_curve,3], texts = Aux_curve, cex=2.5, col="royalblue3" )


 rgl.snapshot(filename = "../Figures/FigSX_Lm_protocol_numbers.png")


#The below technically saves an eps file, but it is way too slow. 
#rgl.postscript("../Figures/Supp_Methods_LM_protocol.pdf", fmt="pdf")





```

#Intraspecific variation - visualizing Macropus rufus specimens 
 
```{r}

####~~~~~~~~~~~~Visualizing two specimens of kangaroo with highly divergent shapes ~~~~~~~~~~~~~####


Orufus_M_UNEM9328 <- read.ply("../Data/Raw/plys/Base_scale/Oruf_9328.ply")

#In order to get colour read in, an additional step is required - I don't understand why but it must have to do with the way the plys were generated, because other plys do not have this issue.

Orufus_M_UNEM9328$material$color  <- Orufus_M_UNEM9328$it
Orufus_M_UNEM9328$material$color <- "pink"

Orufus_F_UNEM9332 <- read.ply("../Data/Raw/plys/Base_scale/Oruf_9332f.ply")
Orufus_F_UNEM9332$material$color <- Orufus_F_UNEM9332$it
Orufus_F_UNEM9332$material$color <- "pink"

#Set usrMat for kangaroos - this only needs running once

# FOV = 0 # sets parallel projection
# open3d(); view3d(fov=0);shade3d(Orufus_M_UNEM9328) # open a mesh on rgl
# rooM.dorsal <- par3d()$userMatrix # set by hand, adjust specimen into dorsal positon
# write.csv(rooM.dorsal, "../Figures/Collections_of_sub_images/Intra_spec/rooM.dorsal.csv",row.names=F)
# 
# FOV = 0 # sets parallel projection
# open3d(); view3d(fov=0);shade3d(Orufus_M_UNEM9328) # open a mesh on rgl
# rooM.lateral <- par3d()$userMatrix # set by hand, adjust specimen into dorsal positon
# write.csv(rooM.lateral, "../Figures/Collections_of_sub_images/Intra_spec/rooM.lateral.csv",row.names=F)
# 
# #Set usrMat for female kangaroos - this only needs running once
# FOV = 0 # sets parallel projection
# open3d(); view3d(fov=0);shade3d(Orufus_F_UNEM9332) # open a mesh on rgl
# rooF.dorsal <- par3d()$userMatrix # set by hand, adjust specimen into dorsal positon
# write.csv(rooF.dorsal, "../Figures/Collections_of_sub_images/Intra_spec/rooF.dorsal.csv",row.names=F)
# 
# FOV = 0 # sets parallel projection
# open3d(); view3d(fov=0);shade3d(Orufus_F_UNEM9332) # open a mesh on rgl
# rooF.lateral <- par3d()$userMatrix # set by hand, adjust specimen into dorsal positon
# write.csv(rooF.lateral, "../Figures/Collections_of_sub_images/Intra_spec/rooF.lateral.csv",row.names=F)


#Make screenshots

views <- c("dorsal", "lateral")
sexes <- c("M", "F")
specimens <- c("Orufus_F_UNEM9332", "Orufus_M_UNEM9328")
roo.M.dorsal <- as.matrix(read.csv("../Figures/Collections_of_sub_images/Intra_spec/rooM.dorsal.csv"), header=TRUE)
roo.M.lateral <- as.matrix(read.csv("../Figures/Collections_of_sub_images/Intra_spec/rooM.lateral.csv"), header=TRUE)
roo.F.dorsal <- as.matrix(read.csv("../Figures/Collections_of_sub_images/Intra_spec/rooF.dorsal.csv"), header=TRUE)
roo.F.lateral <- as.matrix(read.csv("../Figures/Collections_of_sub_images/Intra_spec/rooF.lateral.csv"), header=TRUE)


#For the girl:

    
  for(j in views){
   
    
            open3d(FOV=0, userMatrix=as.name(paste("roo.F",j, sep=".")), windowRect=c(0,0,1200,1200), zoom=0.8)
            shade3d(get("Orufus_F_UNEM9332"), alpha=1, lit = T, shininess = 128.0, specular=49)
            options(warn = 0)
            rgl.snapshot(paste("../Figures/Collections_of_sub_images/Intra_spec/",
            "Orufus_F_UNEM9332_",j,".png", sep=""), 
                         fmt="png", top=TRUE)
            rgl.close()
      }
      
#for the boy:

  for(j in views){
   
    
            open3d(FOV=0, userMatrix=as.name(paste("roo.M",j, sep=".")), windowRect=c(0,0,1200,1200), zoom=0.8)
            shade3d(get("Orufus_M_UNEM9328"), alpha=1, lit = T, shininess = 128.0, specular=49)
            options(warn = 0)
            rgl.snapshot(paste("../Figures/Collections_of_sub_images/Intra_spec/",
            "Orufus_M_UNEM9328_",j,".png", sep=""), 
                         fmt="png", top=TRUE)
            rgl.close()
      }



png("../Figures/Supp_Fig_X_Mruf_var.png",width = 1000,  units = "px", pointsize = 16)
layout(mat=matrix(1:6,ncol=3,byrow=F),widths = c(0.5,1,1), 
           heights = rep.int(1, 2))
  par(mar=c(0,0,0,0))
  for(i in specimens){plot.new(); text(0.5, 0.5, i,cex=1.8)}
  for(j in views){
    for(i in specimens){
      tmp <- readPNG(paste("../Figures/Collections_of_sub_images/Intra_spec/", i,"_",j,".png",sep=""))
      plot(c(0,dim(tmp)[2]),c(0,dim(tmp)[1]),type="n", axes=F, xlab = "", ylab = "",asp=T)
      rasterImage(tmp,0,0,dim(tmp)[2],dim(tmp)[1])
    }}
  
  

dev.off()


```


#Intraspecific variation - visualized as PCA plot

```{r}

# PCA Plot of specimens
ind.PCA <- gm.prcomp(ind.coords_Automatic)
summary_ind_PCA <- summary(ind.PCA)

## get species for which ind replicates exist. To be sure this works, the species names have to be factorized.
ind.reps <- names(which(summary(as.factor(ind.SpecData$Species))>1)) 

# preparing axis labels
ind.pc1lab <-paste("PC 1 ","(",
                   round(100*summary_ind_PCA$PC.summary[2,1],1),"%)",sep="")
ind.pc2lab <-paste("PC 2 ","(",
                   round(100*summary_ind_PCA$PC.summary[2,2],1),"%)",sep="")

# Save figure
png("../Figures/FigX_Intrasp_var_PCA.png",width = 1040, height = 800, units = "px", pointsize = 24)  
plot(ind.PCA$x[,1], ind.PCA$x[,2], asp=T, xlab= ind.pc1lab, 
     ylab= ind.pc2lab, pch = 19, cex = 1, col = "grey", bty="l") 
#if text is required, un-hash the below
#text(ind.PCA$x[,2]~ ind.PCA$x[,1], labels=rownames(ind.PCA$x))
# Need to use points() to add on individuals
ind.cols <- sample (rainbow(length(ind.reps)))

for(i in 1:length(ind.reps)){
  
  if(length(which(ind.SpecData$Species == ind.reps[i])) == 2){
    segments(ind.PCA$x[which(ind.SpecData$Species == ind.reps[i])[1],1], 
             ind.PCA$x[which(ind.SpecData$Species == ind.reps[i])[1],2],
             ind.PCA$x[which(ind.SpecData$Species == ind.reps[i])[2],1], 
             ind.PCA$x[which(ind.SpecData$Species == ind.reps[i])[2],2],
             lty=1, lwd = 2)
  }
  if(length(which(ind.SpecData$Species == ind.reps[i])) > 2){
    tmp <- data.frame(ind.PCA$x)
    edge_points <- rownames(tmp[which(ind.SpecData$Species == ind.reps[i]),])[chull(tmp[which(ind.SpecData$Species == ind.reps[i]), c(1,2)])] # find outer edge points
    polygon(tmp[edge_points,c(1,2)], border= "black", lwd=2) # Plot 
  }
  
  points(ind.PCA$x[which(ind.SpecData$Species == ind.reps[i]),1], 
         ind.PCA$x[which(ind.SpecData$Species == ind.reps[i]),2],
         pch=21, bg=ind.cols[i],cex = 1.5) 
  
}
dev.off()



```


#Compare procrustes distances within and between species to check if between-species variation could dominate the dataset


```{r}

species <-ind.SpecData$Species


png("../Figures/FigX_Distance_comparisons.png",width = 1040, height = 800, units = "px", pointsize = 24)


# First calculate all of the pairwise among-specimen differences across the whole sample. Simply:
D.among.specs <- dist(two.d.array(ind.coords_Automatic))
#hist(D.among.specs) # view as a histogram

# Now do it for each specimen only within a species (when there's 2 or more specs)
D.win.specs <- NULL
for(i in unique(species)){
  tmp <- which(species == i)
  if(length(tmp) > 1){
    D.win.specs <- c(D.win.specs, dist(two.d.array(ind.coords_Automatic[,,tmp]))) }
}
#hist(D.win.specs) # view as a histogram
# Now do it between species means 
D.among.spp <- dist(two.d.array(coords_Automatic))
#hist(D.among.spp)
# Look at them on the same graph
dens.1<-density(D.among.specs)
dens.2<-density(D.win.specs)
dens.3<-density(D.among.spp)
plot(range(dens.1$x, dens.2$x, dens.3$x), range(dens.1$y, dens.2$y, dens.3$y), type = "n", xlab = "Procrustes distance",
     ylab = "Density",xlim = c(0.0, 0.35), ylim = c(0,40), bty="n")
polygon(dens.1, col=rgb(0, 0, 1,0.5), border="black") # among all specimens in dataset
polygon(dens.2, col=rgb(0, 1, 0,0.5), border="black") # within species
polygon(dens.3, col=rgb(1, 0, 0,0.5), border="black") # among species means
legend(0.20, 25, legend = c("among all specimens", "within species", "among species means"),
       pch=21, pt.bg= c(rgb(0, 0, 1,0.5), rgb(0, 1, 0,0.5), rgb(1, 0, 0,0.5)))

dev.off()

remove(tmp, D.win.specs, D.among.specs, D.among.spp, dens.1, dens.2, dens.3, i)

```


# Warp the refmesh to the min and max of PC 1 and PC 2 and save mesh. ONLY NEEDS RUNNING ONCE, takes a while


```{r}
# #With geomorph 3 and the re-arrangement of the PCA code, the below lovely loop sadly doesn't work any more. So doing it manually here.
# 
# PC1min_ply <- plotRefToTarget(ref, PCA$shapes$shapes.comp1$min, method="surface", mesh = refmesh, mag=1)
# open3d()
# shade3d(PC1min_ply)
# writePLY("../Figures/3DModels/PC_warps/PC1min.ply",  withColors = T,  format = "ascii" )
#  
# PC1max_ply <- plotRefToTarget(ref, PCA$shapes$shapes.comp1$max, method="surface", mesh = refmesh, mag=1)
# open3d()
# shade3d(PC1max_ply)
# writePLY("../Figures/3DModels/PC_warps/PC1max.ply",  withColors = T,  format = "ascii" )
# 
# PC2min_ply <- plotRefToTarget(ref, PCA$shapes$shapes.comp2$min, method="surface", mesh = refmesh, mag=1)
# open3d()
# shade3d(PC2min_ply)
# writePLY("../Figures/3DModels/PC_warps/PC2min.ply",  withColors = T,  format = "ascii" )
#  
# PC2max_ply <- plotRefToTarget(ref, PCA$shapes$shapes.comp2$max, method="surface", mesh = refmesh, mag=1)
# open3d()
# shade3d(PC2max_ply)
# writePLY("../Figures/3DModels/PC_warps/PC2max.ply",  withColors = T,  format = "ascii" )
# 
# 
# 
# # 
# #  we can call each PC shapes from this list (PCA made in analyses)
# attach(PCA$pc.shapes)
# 
# PCs <- c("PC1min", "PC1max", "PC2min", "PC2max")
# for (i in PCs) {
#   tmp <-
#     plotRefToTarget(ref,
#                     get(i),
#                     method = "surface",
#                     mesh = refmesh,
#                     mag = 1)
#   assign(paste(i, sep = ""), tmp)
#   
#   open3d()
#   shade3d(tmp)
#   writePLY(
#     paste("../Figures/3DModels/PC_warps/", i, ".ply", sep = ""),
#     withColors = T,
#     format = "ascii"
#   )
#   options(warn = 0)
# } # Ply meshes saved.


```

#After the PC meshes have been created, they can just be loaded as below

```{r}

remove(PC1min_ply,PC1max_ply,PC2min_ply,PC2max_ply)

filelist <- list.files("../Figures/3DModels/PC_warps")
read.ply("../Figures/3DModels/PC_warps/PC2min.ply")


#Read in the warps. The "assign" line involves chopping the .ply file extension off the mesh
for (i in (1:length(filelist))){
  
  tmp<-read.ply(paste("../Figures/3DModels/PC_warps/",filelist[i], sep = ""))
  
  tmp$material$color <- tmp$it
  
  tmp$material$color <- "pink" 
 
  assign(paste( sub(".ply", "_mesh",filelist[i]),sep=""), tmp)
  
}





```

# Make the warp movies

```{r}

#This can't be turned into a loop because of the user interaction, but it's quick. Make sure to maximize the window because it literally takes screenshots, so small window sizes -> low res movie. The resulting images were exported into Fiji as an image stack and then turned into a move. It is worth making the file names so that they are alphabetically orderable in the way they are to be displayed (e.g. 1_warpmovie_PC1_lateral, 2_warpmovie_PC1_dorsal...)

warpmovie3d(PC1max_mesh, PC1min_mesh, n=20, palindrome = TRUE,
            folder = "../Figures/3DModels/Warpmovies/PC1", movie = "warpmovie_PC1_posterior", col = "pink", add = FALSE, close = TRUE,countbegin = 0, ask = TRUE)

warpmovie3d(PC2max_mesh, PC2min_mesh, n=20, palindrome = TRUE,
            folder = "../Figures/3DModels/Warpmovies/PC2", movie = "warpmovie_PC2_posterior", col = "pink", add = FALSE, close = TRUE, countbegin = 0, ask = TRUE)

```


## Plotting meshes and saving a screenshot of each This loop saves all PCs and views to wd/Figures/Fig_parts

```{r}
views <- c("dorsal", "lateral", "posterior")
PCs <- c("PC1min", "PC1max", "PC2min", "PC2max")
for(i in PCs){
  for(j in views){
    open3d(FOV=0, userMatrix=as.name(paste("usrMat", j, sep=".")), windowRect=c(0,0,1200,1200))
    shade3d(get(paste(i,"_mesh",sep="")), alpha=1, lit = T, shininess = 128.0, specular=49)
    options(warn = 0)
    rgl.snapshot(paste("../Figures/Collections_of_sub_images/","Mesh_",i,"_",j,".png", sep=""), 
                 fmt="png", top=TRUE)
    rgl.close()
  }} 
# Finally, trim the white space from those figures using ImageMagick
#system calls ImageMagick and runs a commandline in ImageMagick. Note that this is a command line run inside of Image Magick, using the mogrify -trim functions.

system("magick mogrify C:/Workspace/Github_Projects/Brain_shape/Figures/Collections_of_sub_images/Mesh_*.png -trim C:/Workspace/Github_Projects/Weisbecker_et_al_Brain_shape/Figures/Collections_of_sub_images/*.png")
# These images are used for plots later on


```

#make mean shapes for each clade

```{r}
#using table() to find those clade names that have more than 1 entry 

Clades_nonsingles <- c(names(which((table(gdf_automatic$SpecData.Cladename)>1) == TRUE)))

#Now make mean shape for each of the clades. I admit to manually checking if this really works ;-)

temp=NULL

for (i in c(1:length(Clades_nonsingles))) {

temp  <-  mshape(gdf_automatic$coords[,,which(gdf_automatic$SpecData.Cladename == Clades_nonsingles[i])])

assign( paste(Clades_nonsingles[i], "_mshape", sep=""), temp, )

}

#now warp the mean shape of all to the mean shape of each clade

temp=NULL
Clade_mshapes <- paste(Clades_nonsingles, "_mshape", sep = "")

for (i in Clade_mshapes) {

temp <- plotRefToTarget(ref, get(i), method = "surface", mesh = refmesh, mag=1)

open3d()

shade3d(temp)

writePLY(paste("../Figures/3DModels/Mean_clade_shapes/", i,".ply", sep = ""),  withColors = T,  format = "ascii" )


}

####Get snapshots of all brains after the above has been run

views <- c("dorsal", "lateral", "posterior")

for (i in Clades_nonsingles){
  
    temp <- read.ply(paste("../Figures/3DModels/Mean_clade_shapes/", i,"_mshape",".ply", sep = ""), ShowSpecimen = FALSE)
  temp$material$color <- temp$it
  temp$material$color <- cols[i]
    
    for (j in views) {
  open3d(FOV=0, userMatrix=as.name(paste("usrMat", j, sep=".")), windowRect=c(0,0,1200,1200), zoom=0.9)
  shade3d(temp)
  
  rgl.snapshot(paste("../Figures/Collections_of_sub_images/CladeShapes/",i,"_mshape","_",j,".png", sep=""), 
                         fmt="png", top=TRUE)
            rgl.close()
                }
                    }

#Pick up a few notable brains. Very shoddily using the original ursmats which give a lateral shot of all specimens here.

Additional_species <- c("Ntyph_AMNH202107", "Tcyn_AMNH35244", "Tcarn_SAM280507", "Apyg_J13748", "Pxan_AUM40003", "Cfulig_KU124015", )

cols_add_sp <- c("#0099FF", "dodgerblue2", "#FFCC00", "#FF6600", "#CC0000","light green")
names(cols_add_sp) <-Additional_species

for (i in Additional_species){
  
  temp <- read.ply(paste("../Data/Raw/plys/", i,".ply", sep = ""), ShowSpecimen = FALSE)
  temp$material$color <- temp$it
  temp$material$color <- cols_add_sp[i]
  
  open3d(FOV = 0, userMatrix = usrMat.posterior, windowRect = c(0, 0, 1200, 1200), zoom = 0.8)
  shade3d(temp)
  
  rgl.snapshot(paste("../Figures/Collections_of_sub_images/CladeShapes/",i,".png", sep=""), 
                         fmt="png", top=TRUE)
            rgl.close()
  
}

#Silvabestius and Planigale needs to go separately bc of the rotation


Sjohn <- read.ply(paste("../Data/Raw/plys/Sjohn_QMF30504.ply"), ShowSpecimen = FALSE)
  Sjohn$material$color <- Sjohn$it
  Sjohn$material$color <- "#FFCC00"
  
  open3d(FOV = 0, windowRect = c(0, 0, 1200, 1200), zoom = 0.8)
  #rotate now
  shade3d(Sjohn)
  
  rgl.snapshot(paste("../Figures/Collections_of_sub_images/CladeShapes/","Sjohn_QMF30504",".png", sep=""), 
                         fmt="png", top=TRUE)
            rgl.close()
            
            
  Ping <- read.ply(paste("../Data/Raw/plys/Ping_JM4943.ply"), ShowSpecimen = FALSE)
  Ping$material$color <- Ping$it
  Ping$material$color <- "blue"
  
  open3d(FOV = 0, windowRect = c(0, 0, 1200, 1200), zoom = 0.8)
  #rotate now
  shade3d(Ping)
  
  rgl.snapshot(paste("../Figures/Collections_of_sub_images/CladeShapes/","Ping_JM4943",".png", sep=""), 
                         fmt="png", top=TRUE)
            rgl.close()        
            
        
```
#Figure with phylogeny and warped brains

```{r}

Plot_tree <- trees[3]

pdf(file = "../Figures/Collections_of_sub_images/CladeShapes/Phylo.pdf", height = 12, width=10)

plot(Plot_tree, type="fan", cex=0.9, tip.color = col.gp)

dev.off()


```


Phylomorphospace and locomotion figure preparation - run before making figure just below

```{r}

#This is purely a PCA with one of the trees; only used here

PCA_phylomorpho <- gm.prcomp(gdf_automatic$coords, phy = trees$Tree3, align.to.phy = FALSE)

### Make clade-level colour vector
pch_numbers <- c(24,19, 8, 23, 15, 10, 25)

locomotion <- gdf_automatic$SpecData.Locomotion
locomotion [which(is.na(locomotion)==TRUE)] <- "Extinct"
names(pch_numbers) =  levels(as.factor(locomotion))
pch_loco <- pch_numbers[match(locomotion, names(pch_numbers))]   

```

#PCA Figure

```{r}
png("../Figures/FigX_PCA.png",width = 1040, height = 1400, units = "px", pointsize = 24)

mat <- matrix(c(1,1,2:5), nrow=3,byrow=T)

layout(mat, widths=c(1,1,0.5,0.5), heights=c(0.55,0.3,0.3))# set the size of the rows and columns
#Plot PC1 vs PC2
par(mar=c(4, 5, 0, 1))  # sets the margins c(bottom, left, top, right)


plot(PCA_phylomorpho, phylo = TRUE, pch = pch_loco, cex = 1.5,  cex.axis=1.3, cex.lab=1.3, col = col.gp, bg = col.gp, phylo.par = list(edge.color="black", node.cex=NULL,  tip.labels = FALSE, node.labels = FALSE, edge.color = "gray10"),bty="n",asp=T, ylim = c(-0.1, 0.08) )


#Locomotion legend

legend(x=0.06,y=-0.09,
       legend=names(pch_numbers),
       pch = pch_numbers,
       col="black",
       bg="black",
       bty = "n",
       ncol=2,
       cex = 1.1,
       y.intersp = 0.8)

# Add coloured polygons for the following clades:
clades <- levels(as.factor(SpecData$Cladename))[c(1,2,4,7,11)]
for(i in clades){
  #remove "& SpecData$Fossil == "no" from the below to get Supplementary Fig
  tmp <- which(SpecData$Cladename == i & SpecData$Fossil == "no")
  EP <- rownames(PCA$x[tmp,])[chull(PCA$x[tmp, c(1,2)])] 
  polygon(PCA$x[EP,c(1,2)], col=adjustcolor(cols[which(names(cols) == i)],
                                                    alpha.f = 0.1), border=cols[which(names(cols) == i)]) 
}
# the "Possums"
poss <- c(which(SpecData$Cladename == "Petauroid"),
          which(SpecData$Cladename == "Phalangeroid"))
EP <- rownames(PCA$x[poss,])[chull(PCA$x[poss, c(1,2)])] 
polygon(PCA$x[EP,c(1,2)], col=adjustcolor(cols[which(names(cols) == "Phalangeroid")], alpha.f = 0.1), border=cols[which(names(cols) == "Phalangeroid")]) 
remove(EP, tmp) # clean up environment

#Add species or clade names

text(0.22,0.035, labels = expression (italic("N. typhlops")), col ="#0099FF" )
text(0.14,0.016, labels = expression (italic("C. fuliginosus")), col ="light green" )
text(-0.05, 0.07, labels = "Macropodiformes", col = "#CC0000" )
text(0.065, 0.045, labels = "Petauroid/Phalangeroid possums", col = "#FF6600" )
text(0.1, 0.025, labels = "Peramelemorphia", col = "#9966FF" )
text(0.11, -0.07, labels = "Dasyuromorphia", col = "blue" )
text(-0.09, -0.04, labels = "Didelphimorphia", col = "dark green" )
text(-0.07, -0.016, labels = "Vombatiformes", col = "#FFCC00" )
text(-0.122, 0.01, labels = expression (italic("B. nalima")), col = "#CC0000" )
text(-0.13, -0.0035, labels = expression (italic("E.ima")), col = "#CC0000" )
lines(c(-0.0689497454, -0.105), c(0.0037306146, 0.009), col = "#CC0000" ) # Balbaroo
lines(c (-0.0589188992, -0.12), c(-0.0033289847, -0.0055),col = "#CC0000") #E.ima

#This was not done in a loop because the margins needed differential setting so all the brain images appear the same size (smaller mar numbers mean larger representation of the brain.

#Plotting PC1min
par(mar=c(0.5,0.5,0.5,0.5)) 
tmp <- readPNG(paste("../Figures/Collections_of_sub_images/Mesh_PC/Mesh_PC",1,"min_lateral.png",sep=""))
  plot(c(0,dim(tmp)[2]),c(0,dim(tmp)[1]),type="n", axes=F, xlab = "", ylab = "",asp=T)
  rasterImage(tmp,0,0,dim(tmp)[2],dim(tmp)[1])
  text(60,10, labels=paste("PC",1," min.", sep=""),cex=1.3)

# Plotting PC1max
par(mar=c(3,3,3,3))
tmp <- readPNG(paste("../Figures/Collections_of_sub_images/Mesh_PC/Mesh_PC",1,"max_lateral.png",sep=""))
  plot(c(0,dim(tmp)[2]),c(0,dim(tmp)[1]),type="n", axes=F, xlab = "", ylab = "",asp=T)
  rasterImage(tmp,0,0,dim(tmp)[2],dim(tmp)[1])
  text(60,10, labels=paste("PC",1," max.",sep=""),cex=1.3)
  
  
  #Plotting PC2min
par(mar=c(0.5,0.5,0.5,0.5)) 
tmp <- readPNG(paste("../Figures/Collections_of_sub_images/Mesh_PC/Mesh_PC",2,"min_lateral.png",sep=""))
  plot(c(0,dim(tmp)[2]),c(0,dim(tmp)[1]),type="n", axes=F, xlab = "", ylab = "",asp=T)
  rasterImage(tmp,0,0,dim(tmp)[2],dim(tmp)[1])
  text(60,10, labels=paste("PC",2," min.", sep=""),cex=1.3)

# PlottingPC2max
par(mar=c(1,1,1,1))
tmp <- readPNG(paste("../Figures/Collections_of_sub_images/Mesh_PC/Mesh_PC",2,"max_lateral.png",sep=""))
  plot(c(0,dim(tmp)[2]),c(0,dim(tmp)[1]),type="n", axes=F, xlab = "", ylab = "",asp=T)
  rasterImage(tmp,0,0,dim(tmp)[2],dim(tmp)[1])
  text(60,10, labels=paste("PC",2," max.",sep=""),cex=1.3)
  
dev.off() # end writing png file


```

#Figure SX - Lollipops: Plotting PC shape changes as "LOLLIPOP" vector graphs in order to see subtle shape differences _ this block of code is to make the graphs we used to write the results sections.

```{r}

## Plotting lollipop graphs, saving a screenshot of each
## This loop saves all PCs and views to wd/Figures/Fig_parts!!
# This is bloody amazing, because it save hours of work!
#windowRect gives the size of the window, increase if the image is not fully on the page




views <- c("dorsal", "lateral")
PCs <- c("PC1min", "PC1max", "PC2min", "PC2max")
for(i in PCs){
  for(j in views){
    open3d(FOV=0, userMatrix=as.name(paste("usrMat", j, sep=".")), 
           windowRect=c(0,0,1800,1800), zoom=0.6)
    plotRefToTarget(ref, get(i),method="vector",mag=1, axes=F, gridPars = lmk)
    options(warn = 0)
    rgl.snapshot(paste("../Figures/Collections_of_sub_images/","Lollipop_",i,"_",j,".png", sep=""), 
                 fmt="png", top=TRUE)
    rgl.close()
  }} 

png("../Figures/Supp_Fig_X_Lollipops.png",width = 1800, height = 3000, units = "px", pointsize = 16)
  layout(mat=matrix(1:12,ncol=3,byrow=F),widths = c(0.5,1,1), 
           heights = rep.int(1, 6))
  par(mar=c(0,0,0,0))
  for(i in PCs){plot.new(); text(0.5, 0.5, i,cex=6)}
  for(j in views){
    for(i in PCs){
      tmp <- readPNG(paste("../Figures/Collections_of_sub_images/Lollipop_",i,"_",j,".png",sep=""))
      plot(c(0,dim(tmp)[2]),c(0,dim(tmp)[1]),type="n", axes=F, xlab = "", ylab = "",asp=T)
      rasterImage(tmp,0,0,dim(tmp)[2],dim(tmp)[1])
    }}
dev.off()

```

#Does the bipedality of kangaroos result in a re-arrangement of the brain stem area as a major difference between kangaroos and the other diprotodontia? This uses the heatplot function by landvaR (Weisbecker et al. 2019)

```{r}


#Coordinates of Macropodiformes and (Diprotodontia without Macropodiformes)

MacropCoord <- gdf_automatic$coords[,,which(gdf_automatic$SpecData.Macropod=="yes")]

DipCoord <- gdf_automatic$coords[,,which( gdf_automatic$SpecData.Clade_major =="Diprotodontia" & gdf_automatic$SpecData.Macropod=="no" )]

#creating mean shape for Diprotodontia minus Macropodiforms, and Macropodiforms

Macrop_mshape <- mshape(MacropCoord)
Dip_mshape <- mshape(DipCoord)
  

#find difference between mshapes, used to colourize the heat plot below
difference_DipRoo <- coordinates.difference (mshape(DipCoord), mshape(MacropCoord), type = "spherical")[[1]]


gridPar = gridPar(pt.bg = "white", pt.size = 0.5)


  for(i in views){
    open3d(FOV=0, userMatrix=as.name(paste("usrMat", i, sep=".")), windowRect=c(0,0,1200,1200))
    procrustes.var.plot( Dip_mshape,
                     Macrop_mshape,
                     col = heat.colors, pt.size = 0.8, col.val = difference_DipRoo[,1]  )
    options(warn = 0)
    rgl.snapshot(paste("../Figures/Collections_of_sub_images/","Heat_",i,".png", sep=""), fmt="png", top=TRUE)
    rgl.close()
  } 

system("magick mogrify C:/Workspace/Github_Projects/Brain_shape/Figures/Collections_of_sub_images/Heat_*.png -trim C:/Workspace/Github_Projects/Weisbecker_et_al_Brain_shape/Figures/Collections_of_sub_images/*.png")




mfrow3d(1,2)

 procrustes.var.plot( Dip_mshape,
                     Macrop_mshape,
                     col = heat.colors, pt.size = 0.8, col.val = difference_DipRoo[,1] )

next3d()
 procrustes.var.plot( Dip_mshape,
                     Macrop_mshape,
                     col = heat.colors, pt.size = 0.8, col.val = difference_DipRoo[,1]  )

rgl.snapshot("../Figures/Collections_of_sub_images/SuppX_Heat.png")


#Try as I might, this will not export automatically in satisfactory resolution - therefore I'm just exporting it from the 
#png("../Figures/SuppX_heat.png",width = 1040, height = 1200, units = "px", pointsize = 3, res=1200)
 
 mat <- matrix(c(1,1))

Heat <-readPNG("../Figures/Collections_of_sub_images/SuppX_Heat.png")
   plot(c(0,dim(Heat)[2]),c(0,dim(Heat)[1]),type="n", axes=F, xlab = "", ylab = "",asp=T)
  rasterImage(Heat,0,0,dim(Heat)[2],dim(Heat)[1])
  text(600,200, labels="lateral view, top is dorsal;",cex=1)
    text(1900,200, labels="dorsal view, right is anterior",cex=1)

#dev.off()




```



#Evolutionary Allometry - preparation

```{r}


#For visualisation: allometry without phylogeny

Allom <- procD.lm(coords_Automatic ~ Csize_Automatic,iter=1000, data = gdf_automatic, logsz = TRUE)

Allom_plot <- plot(Allom, type = "regression", predictor = gdf_automatic$Csize, reg.type = "RegScore") 

preds <- shape.predictor(Allom$GM$fitted, x= Allom_plot$RegScore,
                         predmin =  min (Allom_plot$RegScore),
                         predmax= max (Allom_plot$RegScore) 
                         )        

plotRefToTarget(Automatic_GPA$consensus, preds$predmin, method = "vector")

```

#Warping meshes ONLY RUN ONCE - AFTERWARDS, USE THE BELOW READING-IN CODE
```{R}

open3d()

MinCsize <- plotRefToTarget(Automatic_GPA$consensus, preds$predmin, method = "surface", mesh = refmesh, mag=1)
shade3d(MinCsize)
writePLY("../Figures/3DModels/Allom_warps/MinCsize_mesh.ply",withColors=T, format = "ascii")

open3d()

MaxCsize <- plotRefToTarget(Automatic_GPA$consensus, preds$predmax, method = "surface", mesh = refmesh, mag=1)
shade3d(MaxCsize)  
writePLY("../Figures/3DModels/Allom_warps/MaxCsize_mesh.ply",withColors=T, format="ascii")
```


######OR read in the meshes
```{r}
AllomPlys <- c("MinCsize_mesh", "MaxCsize_mesh")

for (i in AllomPlys){

  tmp <- read.ply(paste("../Figures/3DModels/Allom_warps/", i,".ply", sep = "") )
  tmp$material$color <- refmesh$it
  tmp$material$color <- "pink"
  assign (paste(i), tmp )

}

```

#Below are allometry warps but these are just for general interest (not reported)

```{r}
  
## Again using the loop to save all images 

views <- c("dorsal", "lateral", "posterior")
sizes <- c("MinCsize_mesh", "MaxCsize_mesh")

for(i in sizes){
  for(j in views){
    open3d(FOV=0, userMatrix=as.name(paste("usrMat", j, sep=".")), windowRect=c(0,0,1200,1200))
    shade3d(get(paste(i,sep="")), alpha=1, lit = T, shininess = 128.0, specular=49)
    options(warn = 0)
    rgl.snapshot(paste("../Figures/Collections_of_sub_images/Allometry/","Mesh_",i,"_",j,".png", sep=""), fmt="png", top=TRUE)
    rgl.close()
  }} 

system("magick mogrify C:/Workspace/Github_Projects/Brain_shape/Figures/Collections_of_sub_images/Mesh_*.png -trim C:/Workspace/Github_Projects/Weisbecker_et_al_Brain_shape/Figures/Collections_of_sub_images/*.png")


```

#ALLOMETRY FIGURE - Lollipop graphs comparing shape variation between PC1 and Allometry - @@@ideally I'd like this under the allometry figure

```{r}


open3d( userMatrix= usrMat.dorsal,windowRect=c(0,0,1400,700), zoom=0.8 )

mfrow3d(1,2)

plotRefToTarget(preds$predmax,preds$predmin, method="vector", gridPars=lmk)

next3d()
plotRefToTarget(PCA$shapes$shapes.comp1$min, PCA$shapes$shapes.comp1$max, method="vector", gridPars=lmk)

rgl.snapshot("../Figures/Collections_of_sub_images/PC1_Allom_dors.png")
#fmt="png", top=TRUE)

system("magick mogrify C:/Workspace/Github_Projects/Brain_shape/Figures/Collections_of_sub_images/PC1_Allom_dors.png -trim C:/Workspace/Github_Projects/Weisbecker_et_al_Brain_shape/Figures/Collections_of_sub_images/*.png")


```

#Evolutionary allometry with convex hulls of groups that don't fit the common pattern. 

```{r}

# adjust for the fact that procD.allometry reorders the data...!


reorder <- match(dimnames(coords_Automatic)[[3]],row.names(Allom$data)) 



# Save figure
png("../Figures/FigX_allom_PC.png",width = 1040, height = 1200, units = "px", pointsize = 24)
 
mat <- matrix(c(1,2), nrow=2,byrow=T)

#making the top row higher than the bottom row
layout(mat,heights= c(1,0.75))# set the size of the rows and columns

# c(bottom, left, top, right)
par(mar=c(4, 5, 0, 1))


 plot(log(Allom$data$Csize)[reorder], Allom_plot$RegScore[reorder], pch = fos.gp, col=col.gp, bg=col.gp, 
       cex=1.5, xlab = "log(Centroid size)", ylab= "Regression Score")
 text( 4.5,-0.12, labels = "Peramelemorphia", col = "#9966FF" ) 
 text( 5.5,-0.05, labels = "Vombatiformes", col = "#FFCC00" )
  #if checks are required
  #text(Allom$data$Y[reorder]~log(Allom$data$Csize)[reorder], labels=row.names(Allom$data)[reorder] )
  
  
  tmp <- data.frame(log(Allom$data$Csize)[reorder], Allom_plot$RegScore[reorder])
  # Vombatiform
  edge_points <- rownames(tmp[which(SpecData$Cladename == "Vombatiform"),])[chull(tmp[which(SpecData$Cladename == "Vombatiform"), c(1,2)])] # find outer edge points
  polygon(tmp[edge_points,c(1,2)], col=adjustcolor(cols[which(names(cols) == "Vombatiform")], 
          alpha.f = 0.3), border=cols[which(names(cols) == "Vombatiform")]) # Plot polygon
  # Peramelemorph
  edge_points <- rownames(tmp[which(SpecData$Cladename == "Peramelemorph"),])[chull(tmp[which(SpecData$Cladename == "Peramelemorph"), c(1,2)])] # find outer edge points
  polygon(tmp[edge_points,c(1,2)], col=adjustcolor(cols[which(names(cols) == "Peramelemorph")], 
          alpha.f = 0.3), border=cols[which(names(cols) == "Peramelemorph")]) # Plot polygon
  remove(tmp)
  

  
  
  

#Commenting this out because the legend is already in the PCA graph 
  
   # legend("bottomleft", 
  #       
  # legend=c("Dasyurom.","Didelphim.", expression(italic("Dromiciops")), 
  #        "Macropodoid", expression(italic("Notoryctes")),
  #        expression(italic("Caenolestes")), "Peramelem.", 
  #        "Possums", "stem", 
  #        "Vombatiform",expression(italic("Yalkaparidon")),"extinct"), 
  # pt.bg=c(unique(aggregate(col.gp ~ SpecData$Cladename, FUN=unique)[,-1]),
  #       "light grey"), pch=c(rep(21,11),24), cex=1, pt.cex=1, ncol = 3, 
  # bg= "white", box.col="white", text.width = 0.52, y.intersp = 1)
  # 
  
  par(mar=c(0,1,0,0))
  
 Allom_lolli <-readPNG("../Figures/Collections_of_sub_images/PC1_Allom_dors.png")
   plot(c(0,dim(Allom_lolli)[2]),c(0,dim(Allom_lolli)[1]),type="n", axes=F, xlab = "", ylab = "",asp=T)
  rasterImage(Allom_lolli,0,0,dim(Allom_lolli)[2],dim(Allom_lolli)[1])
  text(400,50, labels="small vs. large endocasts",cex=1)
  
  text(1100,50, labels="low vs. high PC1 scores",cex=1)
  
  dev.off()


  #png("../Figures/FigureX_Allom_Pc.png",width = 1040, height = 1200, units = "px", pointsize = 24)

```
#Figure preparing volume and shape PCA comparisons

```{r}

 
# preparing axis labels
pc1lab.vol <-paste("PC 1 ","(", 
                   round(100*PCA.vol$sdev[1],1),"%)",sep="")
pc2lab.vol <-paste("PC 2 ","(",
                   round(100*PCA.vol$sdev[2],1),"%)",sep="")
pc3lab.vol <-paste("PC 3 ","(",
                   round(100*PCA.vol$sdev[3],1),"%)",sep="")

# PC1 vs PC2 with loadings
# Save figure
png("../Figures/FigX_Volumes.png",width = 1040, height = 1100, units = "px", pointsize = 24)
  mat <- matrix(c(1,1,2,3,4,4), nrow=3, ncol=2, byrow = T)
  layout(mat, widths=c(1,1), heights=c(1,0.4,0.25))# set the size of the rows and columns
  #plot 1 - PCA
  par(mar=c(4, 5, 2, 2))  # sets the margins c(bottom, left, top, right)
  plot(PCA.vol$scores[,1],PCA.vol$scores[,2], asp=T, xlab= pc1lab.vol, ylab= pc2lab.vol, 
       pch = fos.gp, cex = 1.5, col=col.gp, bg = col.gp) # plots PCA
  
  # Add coloured polygons for the following clades (including fossils)
clades <- levels(as.factor(SpecData$Cladename))[c(1,2,4,7,11)]
for(i in clades){
  #remove "& SpecData$Fossil == "no" from the below to get Supplementary Fig
  tmp <- which(SpecData$Cladename == i & SpecData$Fossil == "no")
  EP <- rownames(PCA.vol$scores[tmp,])[chull(PCA.vol$scores[tmp, c(1,2)])] 
  polygon(PCA.vol$scores[EP,c(1,2)], col=adjustcolor(cols[which(names(cols) == i)],
                                                    alpha.f = 0.3), border=cols[which(names(cols) == i)]) 
}

# the "Possums"
poss <- c(which(SpecData$Cladename == "Petauroid"),
          which(SpecData$Cladename == "Phalangerid"))
EP <- rownames(PCA.vol$scores[poss,])[chull(PCA.vol$scores[poss, c(1,2)])] 
polygon(PCA.vol$scores[EP,c(1,2)], col=adjustcolor(cols[which(names(cols) == "Phalangeroid")], alpha.f = 0.3), border=cols[which(names(cols) == "Phalangeroid")]) 

#add clade names

text(0.75,-0.36, labels = expression (italic("N. typhlops")), col ="#0099FF" )
text(-0.4, -0.33, labels = expression (italic("C. fuliginosus")), col ="light green" )
text(-0.7, 0.35, labels = "Macropodiformes", col = "#CC0000" )
text(-0.05, -0.4, labels = "Petauroid/Phalangeroid possums", col = "#FF6600" )
text(0.5, -0.4, labels = "Peramelemorphia", col = "#9966FF" )
text(0.25, 0.4, labels = "Dasyuromorphia", col = "blue" )
text(0.65, 0.4, labels = "Didelphimorphia", col = "dark green" )
text(-0.65, -0.2, labels = "Vombatiformes", col = "#FFCC00" )




#plot 2- bar charts
  par(mar=c(4, 4, 0.05, 1))  # sets the margins c(bottom, left, top, right)
  barplot(PCA.vol$loadings[,1],xlab="PC1 loadings",
          names.arg = c("O.Bulb","Cerebr.","Cerebell.","B.Base")) # plot loadings for PC1
  barplot(PCA.vol$loadings[,2],xlab="PC2 loadings",
          names.arg = c("O.Bulb","Cerebr.","Cerebell.","B.Base")) # plot loadings for PC2
 
  dev.off() 
  


```
